+++ src/test/suite/68k.test.ts
@@ -206,6 +206,11 @@ suite("68k-dis", () => {
 		const dis = print_insn_m68k(insn, 0);
 		assert.strictEqual(dis.text, "moveq #1,d1");
 	});
+	test("divul.l", () => {
+		const insn = new Uint8Array([0x4c, 0x40, 0x00, 0x00]);
+		const dis = print_insn_m68k(insn, 0);
+		assert.strictEqual(dis.text, "divul.l d0,d0,d0");
+	});
 });
 
 suite("68k-fpu", () => {